Reorganized src/lib/db/resume/ files.
