+++ src/app/api/resume/[candidateId]/education/route.ts
@@ -2,7 +2,7 @@
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
 import { getCandidateByCandidateId } from "@/lib/db/resume/candidate"
-import { getEducationByCandidateId } from "@/lib/db/resume/education"
+import { getEducationByCandidateId } from "@/lib/db/resume/education/education"
 
 //
 // GET /api/resume/[candidateId]/education.
