+++ src/app/api/resume/[candidateId]/skillSets/[skillSetId]/route.ts
@@ -2,7 +2,7 @@
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
 import { getCandidateByCandidateId } from "@/lib/db/resume/candidate"
-import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSet"
+import { getSkillSetByCandidateIdAndSkillSetId } from "@/lib/db/resume/skillSets/skillSet"
 
 //
 // GET /api/resume/[candidateId]/skillSets/[skillSetId].
