+++ src/app/api/resume/[candidateId]/skillSets/route.ts
@@ -2,7 +2,7 @@
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
 import { getCandidateByCandidateId } from "@/lib/db/resume/candidate"
-import { getSkillSetsByCandidateId } from "@/lib/db/resume/skillSets"
+import { getSkillSetsByCandidateId } from "@/lib/db/resume/skillSets/skillSets"
 
 // GET request.
 export async function GET(
